calculate overdue fees

+++ BookDelinquentReporter/Services/ILibraryService.cs
@@ -12,6 +12,6 @@ namespace BookDelinquentReporter.Services
         Task<List<Member>> GetAllMembers();
         Task<List<Member>> GetDelinquentMembers();
 
-        float GetAmountOwed(Member m);
+        double GetAmountOwed(Member m);
     }
 }
